+++ CotEditor/Sources/EditorTextView.swift
@@ -185,6 +185,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
         super.font = font
         layoutManager.textFont = font
         layoutManager.usesAntialias = defaults[.shouldAntialias]
+        layoutManager.showsIndentGuides = defaults[.showIndentGuides]
         
         self.ligature = defaults[.ligature] ? .standard : .none
         self.invalidateDefaultParagraphStyle()
@@ -653,7 +654,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
         
         guard self.isEditable else { return super.deleteBackward(sender) }
         
-        if self.multipleDeleteBackward() { return }
+        if self.multipleDelete() { return }
         
         // delete tab
         if self.isAutomaticTabExpansionEnabled,
@@ -1196,6 +1197,20 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
     }
     
     
+    /// whether draws indent guides
+    var showsIndentGuides: Bool {
+        
+        get {
+            return (self.layoutManager as? LayoutManager)?.showsIndentGuides ?? true
+        }
+        
+        set {
+            (self.layoutManager as? LayoutManager)?.showsIndentGuides = newValue
+            self.setNeedsDisplay(self.frame, avoidAdditionalLayout: true)
+        }
+    }
+    
+    
     /// whether text is antialiased
     var usesAntialias: Bool {
         
@@ -1494,8 +1509,9 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
         
         guard diff != 0 else { return }
         
+        let heightPath = (self.layoutOrientation == .vertical) ? \NSSize.width : \NSSize.height
         self.textContainerInset.height = height
-        self.frame.size.height += 2 * diff
+        self.frame.size[keyPath: heightPath] += 2 * diff
         
         // invoke `setToFit()` but only when needed to avoid heavy calculation by large document
         // -> `setToFit()` is required to remove the extra height of the frame that contains a blank margin already
@@ -1627,6 +1643,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
             .checkSpellingAsType,
             .autoLinkDetection,
             .pageGuideColumn,
+            .showIndentGuides,
             .overscrollRate,
             .tabWidth,
             .fontName,
@@ -1686,6 +1703,9 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
                 
                 case .pageGuideColumn:
                     self.setNeedsDisplay(self.frame, avoidAdditionalLayout: true)
+                
+                case .showIndentGuides:
+                    self.showsIndentGuides = new as! Bool
                 
                 case .overscrollRate:
                     self.invalidateOverscrollRate()