+++ CotEditor/Sources/LayoutManager.swift
@@ -67,6 +67,8 @@ final class LayoutManager: NSLayoutManager, InvisibleDrawing, ValidationIgnorabl
     }
     
     var invisiblesColor: NSColor = .disabledControlTextColor
+    
+    var showsIndentGuides = false
     var tabWidth = 0
     
     private(set) var spaceWidth: CGFloat = 0
@@ -135,7 +137,7 @@ final class LayoutManager: NSLayoutManager, InvisibleDrawing, ValidationIgnorabl
             NSGraphicsContext.current?.shouldAntialias = self.usesAntialias
         }
         
-        if UserDefaults.standard[.showIndentGuides] {
+        if self.showsIndentGuides {
             self.drawIndentGuides(forGlyphRange: glyphsToShow, at: origin, color: self.invisiblesColor, tabWidth: self.tabWidth)
         }
         
@@ -349,6 +351,7 @@ private extension NSLayoutManager {
         
         color.set()
         let lineWidth: CGFloat = 0.5
+        let scaleFactor = NSGraphicsContext.current?.cgContext.ctm.a ?? 1
         
         // draw guides logical line by logical line
         for (lineRange, indexes) in indentIndexes {
@@ -367,18 +370,35 @@ private extension NSLayoutManager {
                 
                 return lastLineFragment.maxY - lineFragment.minY
             }()
-            let guideRect = NSRect(x: lineFragment.minX, y: lineFragment.minY, width: lineWidth, height: guideLength).offset(by: origin)
+            let guideSize = NSSize(width: lineWidth, height: guideLength)
             
             // draw lines
             for index in indexes {
                 let glyphIndex = self.glyphIndexForCharacter(at: index)
                 let glyphLocation = self.location(forGlyphAt: glyphIndex)
+                let guideOrigin = lineFragment.origin.offset(by: origin).offsetBy(dx: glyphLocation.x).aligned(scale: scaleFactor)
+                let guideRect = NSRect(origin: guideOrigin, size: guideSize)
                 
-                guideRect.offsetBy(dx: glyphLocation.x, dy: 0).fill()
+                guideRect.fill()
             }
         }
         
         NSGraphicsContext.restoreGraphicsState()
+    }
+    
+}
+
+
+private extension CGPoint {
+    
+    /// Make the point pixel-perfect with the desired scale.
+    ///
+    /// - Parameter scale: The scale factor in which the receiver to be pixel-perfect.
+    /// - Returns: An adjusted point.
+    func aligned(scale: CGFloat = 1) -> Self {
+        
+        return Self(x: round(self.x * scale) / scale,
+                    y: round(self.y * scale) / scale)
     }
     
 }