+++ CotEditor/Sources/NSAttributedString.swift
@@ -27,6 +27,9 @@ import Foundation.NSAttributedString
 
 extension NSAttributedString {
     
+    static let newLine = NSAttributedString(string: "\n")
+    
+    
     /// whole range
     var range: NSRange {
         
@@ -37,8 +40,7 @@ extension NSAttributedString {
     /// concatenate attributed strings
     static func + (lhs: NSAttributedString, rhs: NSAttributedString) -> NSAttributedString {
         
-        let result = NSMutableAttributedString()
-        result.append(lhs)
+        let result = NSMutableAttributedString(attributedString: lhs)
         result.append(rhs)
         
         return result.copy() as! NSAttributedString
@@ -48,11 +50,39 @@ extension NSAttributedString {
     /// concatenate attributed strings
     static func += (lhs: inout NSAttributedString, rhs: NSAttributedString) {
         
-        let result = NSMutableAttributedString()
-        result.append(lhs)
+        let result = NSMutableAttributedString(attributedString: lhs)
         result.append(rhs)
         
         lhs = result.copy() as! NSAttributedString
+    }
+    
+}
+
+
+
+extension Sequence where Self.Element == NSAttributedString {
+    
+    /// Return a new attributed string by concatenating the elements of the sequence, adding the given separator between each element.
+    ///
+    /// - Parameter separator: An attributted string to insert between each of the elements in this sequence.
+    /// - Returns: A single, concatenated attributed string.
+    func joined(separator: Element? = nil) -> Element {
+        
+        let result = NSMutableAttributedString()
+        var iterator = self.makeIterator()
+        
+        if let first = iterator.next() {
+            result.append(first)
+            
+            while let next = iterator.next() {
+                if let separator = separator {
+                    result.append(separator)
+                }
+                result.append(next)
+            }
+        }
+        
+        return result.copy() as! NSAttributedString
     }
     
 }