+++ CotEditor/Sources/TextSelection.swift
@@ -154,37 +154,33 @@ final class TextSelection: NSObject {
     
     
     /// line range (location and length) of the selection (list type)
-    @objc var lineRange: Any? {
+    @objc var lineRange: [Int]? {
         
         get {
-            guard
-                let selectedRange = self.document?.selectedRange,
-                let string = self.document?.string else { return nil }
+            guard let document = self.document else { return nil }
             
-            let startLine = string.lineNumber(at: selectedRange.location)
-            let endLine = string.lineNumber(at: selectedRange.upperBound)
+            let selectedRange = document.selectedRange
+            let string = document.string
             
-            return [startLine,
-                    endLine - startLine + 1]
+            let start = string.lineNumber(at: selectedRange.lowerBound)
+            let end = string.lineNumber(at: selectedRange.upperBound)
+            
+            return [start, end - start + 1]
         }
         
         set {
-            let fuzzyRange: FuzzyRange
-            switch newValue {
-                case let number as Int:
-                    fuzzyRange = FuzzyRange(location: number, length: 1)
-                case let range as [Int] where range.count == 2:
-                    fuzzyRange = FuzzyRange(location: range[0], length: range[1])
-                default:
-                    return
-            }
+            guard
+                let lineRange = newValue,
+                (1...2).contains(lineRange.count),
+                // directly communicate with textView as you can skip line ending conversion for this command.
+                let textView = self.document?.textView
+                else { return }
             
-            // you can ignore actuall line ending type and directly comunicate with textView, as this handle just lines
-            guard let string = self.textView?.string else { return }
+            let fuzzyRange = FuzzyRange(location: lineRange[0], length: lineRange[safe: 1] ?? 1)
             
-            guard let range = string.rangeForLine(in: fuzzyRange) else { return }
+            guard let range = textView.string.rangeForLine(in: fuzzyRange) else { return }
             
-            self.document?.selectedRange = range
+            textView.selectedRange = range
         }
     }