+++ CotEditor/Sources/TokenTextView.swift
@@ -178,10 +178,13 @@ final class TokenTextView: NSTextView {
 
 
 
-extension NSMenu {
+extension TokenRepresentable {
     
-    /// add a menu item to insert variable to TokenTextView
-    func addItems<T: TokenRepresentable>(for variables: [T], target: TokenTextView?) {
+    /// Return a menu item to insert variable to TokenTextView.
+    ///
+    /// - Parameter target: The action target.
+    /// - Returns: A menu item.
+    func insertionMenuItem(target: TokenTextView? = nil) -> NSMenuItem {
         
         let fontSize = NSFont.systemFontSize(for: .small)
         let font = NSFont.menuFont(ofSize: fontSize)
@@ -189,19 +192,19 @@ extension NSMenu {
         paragraphStyle.firstLineHeadIndent = 2 * fontSize
         paragraphStyle.headIndent = 2 * fontSize
         
-        for variable in variables {
-            let token = NSAttributedString(string: variable.token, attributes: [.font: font])
-            let description = NSAttributedString(string: "\n" + variable.localizedDescription, attributes: [.font: font,
-                                                                                                            .foregroundColor: NSColor.gray,
-                                                                                                            .paragraphStyle: paragraphStyle])
-            let item = NSMenuItem()
-            item.target = target
-            item.action = #selector(TokenTextView.insertVariable)
-            item.attributedTitle = token + description
-            item.representedObject = variable.token
-            
-            self.addItem(item)
-        }
+        let token = NSAttributedString(string: self.token, attributes: [.font: font])
+        let description = NSAttributedString(string: self.localizedDescription,
+                                             attributes: [.font: font,
+                                                          .foregroundColor: NSColor.secondaryLabelColor,
+                                                          .paragraphStyle: paragraphStyle])
+        
+        let item = NSMenuItem()
+        item.target = target
+        item.action = #selector(TokenTextView.insertVariable)
+        item.attributedTitle = [token, description].joined(separator: .newLine)
+        item.representedObject = self.token
+        
+        return item
     }
     
 }